Avoid decode of bytes received for WAKE_ACK and ACK tests
On one occasion the decode has triggered a Unicode exception, and it
isn't necessary if the expected WAKE_ACK and ACK values are encoded the
class is instantiated.

Note that the OK constant might have a similar issue, but I wouldn't be
able to test any changes in this case without impacting the weather
statuion in an unacceptable way.