Convert the VantagePro object to use a generic async read/write object
Rather than a VantagePro object that explicitly expects a serial device
name and which creates a Serial object, the code now expects a more
abstracted "device" object capable of async binary read/write.

In the script that uses this new version of davis.py, the object is an
asyncio object based on an IP connection to a specified host name and
port number.

But the PySerial package can be used in conjunction with the
PySerial-asyncio package to allow for async serial interactions. I'm not
certain if PySerial-asyncio has been developed to a point where it would
really work and be functional (no way to test). If it is adequate, then
the PyWeather/weather/stations/davis.py module could be developed to
handle async for either network or serial